pyproject.toml
[project]
name = "spark-medallion-pipeline"
version = "0.1.0"
description = "Batch ETL pipeline with PySpark and Delta Lake following the Medallion architecture"
requires-python = ">=3.11,<3.13"

dependencies = [
    "pyspark==3.5.1",
    "delta-spark==3.1.0",
    "mysql-connector-python>=9.0.0",
    "python-dotenv>=1.0.0",
    "mcp[cli]>=1.0.0",
    "duckdb>=0.10.0",
    "soda-core-spark-df>=3.3.0,<4",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "poethepoet>=0.25.0",
    "pre-commit>=3.7.0",
    "pyright>=1.1.0",
]

[tool.poe.tasks]
linter-fix = "ruff format src"
check = {shell = "ruff check src && pyright src"}
test = "pytest tests/ -v"
coverage = "pytest tests/ --cov=src --cov-report=term-missing"
pipeline = "python src/main.py"
mcp = "python src/mcp_server/server.py"
infra-up = "docker compose up -d"
infra-down = "docker compose down"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"
extraPaths = ["src"]